removed multiplier from DIRECT_HASH_FUNC, not needed with fibonacci hashing DIRECT_HASH_FUNC uses uint64 for 64b types

+++ str.h
@@ -2378,7 +2378,7 @@ template<bool INSENSITIVE> struct hasher<charstr, INSENSITIVE>
 {
     typedef charstr key_type;
 
-    size_t operator() (const charstr& str) const
+    uint operator() (const charstr& str) const
     {
         return INSENSITIVE
             ? __coid_hash_string(str.ptr(), str.len())
